#130 Separated unit test and integration tests for PJMClient.
Also moved expected response data into files in the fixtures/pjm directory.
