+++ tests/fixtures/pjm/ForecastedLoadHistory.html
@@ -1,0 +1,34 @@
+<h1>Forecasted Load History</h1>
+<p>As of <span id="ctl00_ContentPlaceHolder1_DateAndTime">12.11.2015 17:24</span> EDT <a href="javascript:location.reload();">Refresh</a></p>
+<dl class="no-margin">
+    <dt>Name: </dt>
+    <dd>PJM RTO Total</dd>
+</dl>
+<div class="table-tabs">
+    <ul>
+        <li id="tableView" class="selected"><a href="javascript:void(0);">Table View</a></li>
+        <li id="chartView"><a href="javascript:void(0);">Chart View</a></li>
+    </ul>
+    <div id="table-view">
+        <table class="edata-table stripped">
+            <thead>
+            <th>Time Stamp</th>
+            <th class="right">MW</th>
+            </thead>
+            <tbody>
+            <tr id="ctl00_ContentPlaceHolder1_lvForecastedLoadHistoryTable_ctrl0_trRow1">
+                <td>12.11.2015 17:00 EST</td>
+                <td class="right">92,725</td>
+            </tr>
+            <tr id="ctl00_ContentPlaceHolder1_lvForecastedLoadHistoryTable_ctrl1_trRow1">
+                <td>12.11.2015 18:00 EST</td>
+                <td class="right">93,358</td>
+            </tr>
+            <tr id="ctl00_ContentPlaceHolder1_lvForecastedLoadHistoryTable_ctrl2_trRow1">
+                <td>12.11.2015 19:00 EST</td>
+                <td class="right">91,496</td>
+            </tr>
+            </tbody>
+        </table>
+    </div>
+</div>
