+++ tests/fixtures/pjm/InstantaneousLoad.html
@@ -1,0 +1,24 @@
+<h1>Instantaneous Load</h1>
+<p>As of <span id="ctl00_ContentPlaceHolder1_DateAndTime">12.11.2015 17:23</span> EDT <a href="javascript:location.reload();">Refresh</a></p>
+<table class="edata-table stripped">
+    <thead>
+    <tr>
+        <th>Name</th>
+        <th class="right">MW</th>
+    </tr>
+    </thead>
+    <tbody>
+    <tr>
+        <td>PJM RTO Total</td>
+        <td class="right">91,419</td>
+    </tr>
+    <tr>
+        <td>Mid-Atlantic Region</td>
+        <td class="right">33,280</td>
+    </tr>
+    <tr>
+        <td>AE Zone</td>
+        <td class="right">1,212</td>
+    </tr>
+    </tbody>
+</table>
